Fix hybrid search NameError: pass index parameter to _build_hybrid_query method
The hybrid search was failing because _build_hybrid_query method was trying to use
an 'index' variable that wasn't defined in its scope. This caused a NameError when
the method tried to call _get_boosted_fields(index, 'hybrid').

Fixed by:
- Adding index parameter to _build_hybrid_query method signature
- Passing index from hybrid_search method to _build_hybrid_query

This resolves the 'index is not defined' error in hybrid search.